+++ src/main/java/kw/zeropick/config/SwaggerConfig.java
@@ -1,0 +1,26 @@
+package kw.zeropick.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import io.swagger.v3.oas.models.servers.Server;
+
+@Configuration
+public class SwaggerConfig {
+
+    @Value("${api.server.url}")
+    private String serverUrl;
+
+    @Bean
+    public OpenAPI openAPI() {
+        return new OpenAPI()
+                .info(info)
+                .addServersItem(new Server().url(serverUrl));
+    }
+
+    Info info = new Info().title("ZEROPICK Backend APIS").version("0.0.1").description(
+            "<h3>ZEROPICK Backend APIS</h3>");
+
+}
