+++ src/main/java/kw/zeropick/product/domain/Product.java
@@ -5,7 +5,8 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import java.util.ArrayList;
 import java.util.List;
-import kw.zeropick.review.domain.NegativeTag;
+import kw.zeropick.common.converter.StringListToStringConverter;
+import kw.zeropick.review.domain.PositiveTagEnum;
 import kw.zeropick.review.domain.Review;
 import lombok.*;
 
@@ -43,6 +44,9 @@ public class Product {
     private int bookmarkCount;
 
     private int reviewCount;
+
+    @Convert(converter = StringListToStringConverter.class)
+    private List<PositiveTagEnum> tags;
 
     @OneToMany(mappedBy = "product", fetch = FetchType.LAZY)
     private List<Review> reviews = new ArrayList<>();
