--- src/main/java/kw/zeropick/swagger/Config.java
@@ -1,26 +1,0 @@
-package kw.zeropick.swagger;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import io.swagger.v3.oas.models.servers.Server;
-
-@Configuration
-public class Config {
-
-    @Value("${api.server.url}")
-    private String serverUrl;
-
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-                .info(info)
-                .addServersItem(new Server().url(serverUrl));
-    }
-
-    Info info = new Info().title("ZEROPICK Backend APIS").version("0.0.1").description(
-            "<h3>ZEROPICK Backend APIS</h3>");
-
-}
